requirements/local.txt
-r base.txt
coverage==4.5.2
django-coverage-plugin==1.6.0
Sphinx==1.8.3
django-extensions==2.1.4
Werkzeug==0.14.1
django-test-plus==1.1.1
factory_boy==2.11.1
django-debug-toolbar==1.11

# improved REPL
ipdb==0.11

pytest-django==3.4.5
pytest-sugar==0.9.2
